backend: Split batch complete apartment tests to their own file
